test(tbtc): tag signer-material resolver tests to their build flavour
Three signer-material resolver tests were tagged plain `frost_native` but each
asserts behaviour specific to ONE build flavour, so no single build ran them all
green (the failure surfaced when pkg/tbtc was exercised under a non-cgo
frost_native combo):

  - TestRegisterSignerMaterialResolverForBuild_DefaultProviderRefusesScaffoldWithoutOptIn
    asserts the resolver REFUSES scaffold-era material without opt-in. Only the
    cgo `frost_tbtc_signer` resolver carries that guard; the non-cgo
    `frost_native` resolver permits it (transitional build -- the deeper
    native_ffi_primitive guard refuses scaffold material at signing time
    instead). So the test failed under any non-cgo frost_native build.
  - TestUnmarshalSignerMaterialFromPersistence_LegacyEncodingResolvesNativeMaterialOnFrostNativeBuild
  - TestSignerMarshalling_LegacyRoundtripMigratesToNativeEnvelopeOnFrostNativeBuild
    assert legacy->native resolution SUCCEEDS without opt-in -- the permissive
    non-cgo behaviour. So they failed under the cgo combo (resolver refuses).

Re-tag each test to the build whose resolver contract it exercises, mirroring
the resolver files' own tag split (signer_material_resolver_build_frost_native.go
vs signer_material_resolver_build_frost_native_tbtc_signer.go):

  - move the refusal test into a new cgo-tagged file
    (`frost_native && frost_tbtc_signer && cgo`);
  - tag the migration tests `frost_native && !(frost_tbtc_signer && cgo)`.

Test-only change; no production behaviour changes. Verified: gofmt clean; the
signer-material suite is green under frost_native, frost_native+frost_roast_retry,
and frost_native+frost_tbtc_signer (cgo); each moved test executes in its target
build and is excluded from the other.
